requirements.txt
aiohttp
beautifulsoup4
feedparser
readability-lxml
trafilatura
python-telegram-bot

